Hog2 (#30)
* First push of Hog 2.0

* Changed the install dir from hog to hogger

* Changed config file to hogger

* adds embodied carbon

+++ install.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 set -euo pipefail
 
+HOG_PATH="/usr/local/bin/hogger"
+
 GREEN='\033[0;32m'
 NC='\033[0m' # No Color
+CONFIG_FILE="/etc/hogger_settings.ini"
 
 # Function to check and install Xcode Command Line Tools
 install_xcode_clt() {
@@ -34,10 +37,10 @@ install_xcode_clt() {
 install_xcode_clt
 
 # Checks if the hog is already running
-hog_running_output=$(launchctl list | grep io.green-coding.hog || echo "")
+hog_running_output=$(launchctl list | grep io.green-coding.hogger || echo "")
 
 if [[ ! -z "$hog_running_output" ]]; then
-    launchctl unload /Library/LaunchDaemons/io.green-coding.hog.plist
+    launchctl bootout system /Library/LaunchDaemons/io.green-coding.hogger.plist
     rm -f /tmp/latest_release.zip
 fi
 
@@ -48,56 +51,66 @@ fi
 ZIP_LOCATION=$(curl -s https://api.github.com/repos/green-coding-solutions/hog/releases/latest | grep -o 'https://[^"]*/hog_power_logger.zip')
 curl -fLo /tmp/latest_release.zip $ZIP_LOCATION
 
-mkdir -p /usr/local/bin/hog
+if [[ -z "/tmp/latest_release.zip" ]]; then
+    echo "Error: Could not fetch the ZIP URL from the GitHub API."
+    exit 1
+fi
 
-unzip -o -u /tmp/latest_release.zip -d /usr/local/bin/hog/
+mkdir -p "$HOG_PATH"
+
+unzip -o -u /tmp/latest_release.zip -d "$HOG_PATH"
 rm /tmp/latest_release.zip
 
-chmod 755 /usr/local/bin/hog
-chmod -R 755 /usr/local/bin/hog/
-chmod +x /usr/local/bin/hog/power_logger.py
+chmod 755 "$HOG_PATH"
+chmod -R 755 "$HOG_PATH"
+chmod +x "$HOG_PATH"/power_logger.py
 
 ###
 # Writing the config file
 ###
 
-if [[ -t 0 ]]; then  # Check if input is from a terminal
-    read -p "In order for the app to work with all features please allow us to upload some data. [Y/n]: " upload_data
+update_config() {
+    local key="$1"
+    local value="$2"
+    local file="$3"
+    if grep -q "^${key} =" "$file"; then
+        sed -i "s|^${key} =.*|${key} = ${value}|" "$file"
+    else
+        echo "${key} = ${value}" >> "$file"
+    fi
+}
+
+cat settings.ini > "$CONFIG_FILE"
+
+if [[ -t 0 ]]; then
+    read -p "In order for the app to work with all features, please allow us to upload some data. [Y/n]: " upload_data
     upload_data=${upload_data:-Y}
     upload_data=$(echo "$upload_data" | tr '[:upper:]' '[:lower:]')
-
-    if [[ $upload_data == "y" || $upload_data == "yes" ]]; then
-        upload_flag="true"
-    else
-        upload_flag="false"
-    fi
 else
-    upload_flag="true"
+    upload_data="n"
 fi
 
-cat > /etc/hog_settings.ini << EOF
-[DEFAULT]
-api_url = https://api.green-coding.io/v1/hog/add
-web_url = https://metrics.green-coding.io/hog-details.html?machine_uuid=
-upload_delta = 300
-powermetrics = 5000
-upload_data = $upload_flag
-resolve_coalitions=com.googlecode.iterm2,com.apple.Terminal,com.vix.cron
-EOF
+if [[ $upload_data == "y" || $upload_data == "yes" ]]; then
+    sed -i "s|^upload_data =.*|upload_data = true|" "$CONFIG_FILE"
+else
+    sed -i "s|^upload_data =.*|upload_data = false|" "$CONFIG_FILE"
+fi
 
-echo "Configuration written to /etc/hog_settings.ini"
+
+echo "Installation complete. Configuration updated at $CONFIG_FILE."
+
 
 ###
 # Setting up the background demon
 ###
 
-mv -f /usr/local/bin/hog/io.green-coding.hog.plist /Library/LaunchDaemons/io.green-coding.hog.plist
+mv -f "$HOG_PATH/io.green-coding.hogger.plist /Library/LaunchDaemons/io.green-coding.hogger.plist"
 
-sed -i '' "s|PATH_PLEASE_CHANGE|/usr/local/bin/hog|g" /Library/LaunchDaemons/io.green-coding.hog.plist
+sed -i '' "s|PATH_PLEASE_CHANGE|$HOG_PATH|g" /Library/LaunchDaemons/io.green-coding.hogger.plist
 
-chown root:wheel /Library/LaunchDaemons/io.green-coding.hog.plist
-chmod 644 /Library/LaunchDaemons/io.green-coding.hog.plist
+chown root:wheel /Library/LaunchDaemons/io.green-coding.hogger.plist
+chmod 644 /Library/LaunchDaemons/io.green-coding.hogger.plist
 
-launchctl load /Library/LaunchDaemons/io.green-coding.hog.plist
+launchctl bootstrap system /Library/LaunchDaemons/io.green-coding.hogger.plist
 
 echo -e "${GREEN}Successfully installed the Power Hog Demon!${NC}"
